Fix to #1858 - Filter missing if wrong type passed in
Issue is for queries that use Equals(object) method for comparison. C# will return false whenever objects of different types are compared using this method, even if the underlying values are identical after conversion to one of the types.
If arguments are nullable (e.g. int?), nullability is ignored for the purpose of the comparison. E.g. 'int?' and 'int' may result in positive comparison, int? and int? as well, int? and long? will always return false.

Fix is to mimic this behavior, as well as log warning that use of Equals(object) on two objects of different types was probably not intentional.
Another fix is in the way we evaluate parameters. Before we if parameter was wrapped in the Convert node, we would evaluate it's type to the type it was converted to. In case of Equals(object), and probably other cases as well, this causes creation of parameters typed as object, which always get evaluated on the client (since we don't know how to translate parameter/constant of type 'object'.

Fix is to evaluate parameters to it's original type, and wrap the parameter into Convert node afterward. This way we preserver the intended typing but given parameters can be evaluated on the server.